testing rating

+++ src/main/java/org/arQam/AnnonceEfficace/Metier/Utilisateur.java
@@ -80,7 +80,7 @@ public class Utilisateur {
    		return null;
    	}      
 
-	public Utilisateur load(Long id) {
+	public Utilisateur load(int id) {
    	 SessionFactory sf = HibernateUtil.getSessionFactory();
         Session session = sf.openSession();
      
@@ -118,7 +118,7 @@ public class Utilisateur {
 	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getNomUtilisateur() {
